move map and loading file to PokemonList
